Remove with mysql_native_password (#70)
This was mainly needed for PHP 7.1/7.2/7.3 as there was no built-in support for the caching_sha2 authentication method. But it has been added to mysqlnd after 7.4.

As the active support for 7.3 has ended already and soon even the Security fixes will no longer be done I am going to merge this to encourage people to use 7.4 and 8.x.